refactor: change the dir of `useGraphqlMutation` & `useGraphqlQuery` hooks
Move `useGraphqlMutation` and `useGraphqlQuery` hooks from `src/lib/hooks/`
directory into `src/hooks/` directory.
